refactor(go-server): consolidate authorization security files
