+++ packages/server/duckdb-server-go/pkg/server/authorization.go
@@ -7,19 +7,13 @@ import (
 )
 
 var (
-	// ErrUnauthenticated indicates that a request has no valid authentication.
-	ErrUnauthenticated = errors.New("server: unauthenticated")
-	// ErrPermissionDenied indicates that an authenticated request is not allowed
-	// to execute a command.
+	ErrUnauthenticated  = errors.New("server: unauthenticated")
 	ErrPermissionDenied = errors.New("server: permission denied")
-	// ErrInvalidCommand indicates that a command cannot be authorized because it
-	// is malformed or unsupported.
-	ErrInvalidCommand = errors.New("server: invalid command")
+	ErrInvalidCommand   = errors.New("server: invalid command")
 
 	errNilAuthorizerFunc = errors.New("server: nil AuthorizerFunc")
 )
 
-// CommandType identifies the result format or operation requested by a command.
 type CommandType string
 
 const (
@@ -39,12 +33,10 @@ func newCommand(typ CommandType, sql string) Command {
 	return Command{typ: typ, sql: sql}
 }
 
-// Type returns the command's normalized operation type.
 func (c Command) Type() CommandType {
 	return c.typ
 }
 
-// SQL returns the exact SQL text that will be executed.
 func (c Command) SQL() string {
 	return c.sql
 }
@@ -63,11 +55,8 @@ type Authorizer interface {
 	AuthorizeRequest(*http.Request) (CommandAuthorizer, error)
 }
 
-// AuthorizerFunc adapts a function to the Authorizer interface.
 type AuthorizerFunc func(*http.Request) (CommandAuthorizer, error)
 
-// AuthorizeRequest calls f. A nil AuthorizerFunc returns an internal error
-// instead of panicking.
 func (f AuthorizerFunc) AuthorizeRequest(r *http.Request) (CommandAuthorizer, error) {
 	if f == nil {
 		return nil, errNilAuthorizerFunc
